# server/docs/swagger_output_dev_new.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/v1/student/me/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully obtained.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "name": {
                      "type": "string",
                      "example": "John Doe"
                    },
                    "email": {
                      "type": "string",
                      "example": "john@doe.com"
                    },
                    "phone_number": {
                      "type": "string",
                      "example": "+639784561234"
                    },
                    "block": {
                      "type": "string",
                      "example": "A"
                    },
                    "regNo": {
                      "type": "string",
                      "example": "19BCE0000"
                    },
                    "Room": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number",
                          "example": 1
                        },
                        "roomNo": {
                          "type": "string",
                          "example": "101"
                        },
                        "block": {
                          "type": "string",
                          "example": "A"
                        },
                        "studentId": {
                          "type": "number",
                          "example": 1
                        }
                      }
                    },
                    "complaint": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 1
                          },
                          "complaintType": {
                            "type": "string",
                            "example": "Electrical"
                          },
                          "complaintDate": {
                            "type": "string",
                            "example": "2021-05-05"
                          },
                          "complaintDescription": {
                            "type": "string",
                            "example": "Lights not working"
                          },
                          "complaintSeverity": {
                            "type": "string",
                            "example": "High"
                          },
                          "isResolved": {
                            "type": "boolean",
                            "example": false
                          },
                          "studentId": {
                            "type": "number",
                            "example": 1
                          },
                          "wardenId": {
                            "type": "number",
                            "example": 1
                          }
                        }
                      }
                    },
                    "leave": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 1
                          },
                          "leaveType": {
                            "type": "string",
                            "example": "Casual"
                          },
                          "leaveDate": {
                            "type": "string",
                            "example": "2021-05-05"
                          },
                          "leaveTime": {
                            "type": "string",
                            "example": "10:00:00"
                          },
                          "LeaveDuration": {
                            "type": "string",
                            "example": "1"
                          },
                          "isApproved": {
                            "type": "boolean",
                            "example": false
                          },
                          "isRejected": {
                            "type": "boolean",
                            "example": false
                          },
                          "studentId": {
                            "type": "number",
                            "example": 1
                          },
                          "wardenId": {
                            "type": "number",
                            "example": 1
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/v1/student/leave/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully obtained.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "leave": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 1
                          },
                          "leaveType": {
                            "type": "string",
                            "example": "Casual"
                          },
                          "leaveDate": {
                            "type": "string",
                            "example": "2021-05-05"
                          },
                          "leaveTime": {
                            "type": "string",
                            "example": "10:00:00"
                          },
                          "LeaveDuration": {
                            "type": "string",
                            "example": "1"
                          },
                          "isApproved": {
                            "type": "boolean",
                            "example": false
                          },
                          "isRejected": {
                            "type": "boolean",
                            "example": false
                          },
                          "studentId": {
                            "type": "number",
                            "example": 1
                          },
                          "wardenId": {
                            "type": "number",
                            "example": 1
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "leaveType": {
                  "example": "any"
                },
                "leaveDate": {
                  "example": "any"
                },
                "leaveTime": {
                  "example": "any"
                },
                "leaveDuration": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully obtained.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "leaveType": {
                      "type": "string",
                      "example": "Casual"
                    },
                    "leaveDate": {
                      "type": "string",
                      "example": "2021-05-05"
                    },
                    "leaveTime": {
                      "type": "string",
                      "example": "10:00:00"
                    },
                    "LeaveDuration": {
                      "type": "string",
                      "example": "1"
                    },
                    "isApproved": {
                      "type": "boolean",
                      "example": false
                    },
                    "isRejected": {
                      "type": "boolean",
                      "example": false
                    },
                    "studentId": {
                      "type": "number",
                      "example": 1
                    },
                    "wardenId": {
                      "type": "number",
                      "example": 1
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "leaveID": {
                  "example": "any"
                },
                "leaveType": {
                  "example": "any"
                },
                "leaveDate": {
                  "example": "any"
                },
                "leaveTime": {
                  "example": "any"
                },
                "leaveDuration": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully obtained.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "leaveType": {
                      "type": "string",
                      "example": "Casual"
                    },
                    "leaveDate": {
                      "type": "string",
                      "example": "2021-05-05"
                    },
                    "leaveTime": {
                      "type": "string",
                      "example": "10:00:00"
                    },
                    "LeaveDuration": {
                      "type": "string",
                      "example": "1"
                    },
                    "isApproved": {
                      "type": "boolean",
                      "example": false
                    },
                    "isRejected": {
                      "type": "boolean",
                      "example": false
                    },
                    "studentId": {
                      "type": "number",
                      "example": 1
                    },
                    "wardenId": {
                      "type": "number",
                      "example": 1
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "leaveID": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully obtained.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "leaveType": {
                      "type": "string",
                      "example": "Casual"
                    },
                    "leaveDate": {
                      "type": "string",
                      "example": "2021-05-05"
                    },
                    "leaveTime": {
                      "type": "string",
                      "example": "10:00:00"
                    },
                    "LeaveDuration": {
                      "type": "string",
                      "example": "1"
                    },
                    "isApproved": {
                      "type": "boolean",
                      "example": false
                    },
                    "isRejected": {
                      "type": "boolean",
                      "example": false
                    },
                    "studentId": {
                      "type": "number",
                      "example": 1
                    },
                    "wardenId": {
                      "type": "number",
                      "example": 1
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/student/complaint/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully obtained.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "complaint": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 1
                          },
                          "complaintType": {
                            "type": "string",
                            "example": "Electrical"
                          },
                          "complaintDate": {
                            "type": "string",
                            "example": "2021-05-05"
                          },
                          "complaintDescription": {
                            "type": "string",
                            "example": "Lights not working"
                          },
                          "complaintSeverity": {
                            "type": "string",
                            "example": "High"
                          },
                          "isResolved": {
                            "type": "boolean",
                            "example": false
                          },
                          "studentId": {
                            "type": "number",
                            "example": 1
                          },
                          "wardenId": {
                            "type": "number",
                            "example": 1
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "complaintType": {
                  "example": "any"
                },
                "complaintDate": {
                  "example": "any"
                },
                "complaintDescription": {
                  "example": "any"
                },
                "complaintSeverity": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully obtained.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "complaintType": {
                      "type": "string",
                      "example": "Electrical"
                    },
                    "complaintDate": {
                      "type": "string",
                      "example": "2021-05-05"
                    },
                    "complaintDescription": {
                      "type": "string",
                      "example": "Lights not working"
                    },
                    "complaintSeverity": {
                      "type": "string",
                      "example": "High"
                    },
                    "isResolved": {
                      "type": "boolean",
                      "example": false
                    },
                    "studentId": {
                      "type": "number",
                      "example": 1
                    },
                    "wardenId": {
                      "type": "number",
                      "example": 1
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "complaintId": {
                  "example": "any"
                },
                "complaintType": {
                  "example": "any"
                },
                "complaintDate": {
                  "example": "any"
                },
                "complaintDescription": {
                  "example": "any"
                },
                "complaintSeverity": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully obtained.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "complaintType": {
                      "type": "string",
                      "example": "Electrical"
                    },
                    "complaintDate": {
                      "type": "string",
                      "example": "2021-05-05"
                    },
                    "complaintDescription": {
                      "type": "string",
                      "example": "Lights not working"
                    },
                    "complaintSeverity": {
                      "type": "string",
                      "example": "High"
                    },
                    "isResolved": {
                      "type": "boolean",
                      "example": false
                    },
                    "studentId": {
                      "type": "number",
                      "example": 1
                    },
                    "wardenId": {
                      "type": "number",
                      "example": 1
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "complaintId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully obtained.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Deleted successfully"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "complaintType": {
                      "type": "string",
                      "example": "Electrical"
                    },
                    "complaintDate": {
                      "type": "string",
                      "example": "2021-05-05"
                    },
                    "complaintDescription": {
                      "type": "string",
                      "example": "Lights not working"
                    },
                    "complaintSeverity": {
                      "type": "string",
                      "example": "High"
                    },
                    "isResolved": {
                      "type": "boolean",
                      "example": false
                    },
                    "studentId": {
                      "type": "number",
                      "example": 1
                    },
                    "wardenId": {
                      "type": "number",
                      "example": 1
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/student/room-details/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully obtained.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 4
                    },
                    "roomNo": {
                      "type": "string",
                      "example": "A-105"
                    },
                    "roomType": {
                      "type": "string",
                      "example": "AC"
                    },
                    "roomCapacity": {
                      "type": "number",
                      "example": 2
                    },
                    "block": {
                      "type": "string",
                      "example": "A"
                    },
                    "isFull": {
                      "type": "boolean",
                      "example": false
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}